moved htpasswd handling directly into the basicauth provider
