+++ providers/basic/htpasswd_test.go
@@ -1,9 +1,10 @@
-package helpers
+package basic
 
 import (
-	"github.com/stretchr/testify/assert"
 	"io/ioutil"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
 
 const (
